requirements.txt
git+https://github.com/WildDogOne/ediplug-py.git
requests
rich
python-telegram-bot